Update this unmaintained library (#1)
Looks like we forked this library 3 years back and never changed it, but
we import it into message-reply-service. I don't know why we forked it
then if we weren't going to change it.

But now it needs an update, because it no longer works with the recent
versions of other libraries. So we have to maintain this 3d party
unmaintained library.

Here I'm trying to get builds of message-reply-service to work.